# src/main/resources/static/hackhub-openapi.yaml
openapi: 3.0.0
info:
  title: hackhub-openapi
  version: 1.0.0
paths:
...
  /api/invites/accetta?nomeTeam=piedi&utenteId=dadd5008-e361-480e-95b9-89097d364fd5:
    post:
      summary: Accetta un invito
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_inviti_accetta_un_invito_yml
      description: ''
      tags:
        - Inviti
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: nomeTeam
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: piedi
        - name: utenteId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: dadd5008-e361-480e-95b9-89097d364fd5
  /api/invites:
    post:
      summary: Invia un invito a un utente
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_inviti_invia_un_invito_a_un_utente_yml
      description: ''
      tags:
        - Inviti
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      requestBody:
        $ref: '#/components/requestBodies/invia_un_invito_a_un_utente'
  /api/invites/casella/:utenteId:
    get:
      summary: Restituisce la casella degli inviti di un utente
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_inviti_restituisce_la_casella_degli_inviti_di_un_utente_yml
      description: ''
      tags:
        - Inviti
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mittente:
                      type: object
                      properties:
                        nome:
                          type: string
                        maxMembri:
                          type: integer
                        membri:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              nome:
                                type: string
                              cognome:
                                type: string
                              role:
                                type: string
                    destinatario:
                      type: object
                      properties:
                        id:
                          type: string
                        nome:
                          type: string
                        cognome:
                          type: string
                        role:
                          type: string
              example:
                - mittente:
                    nome: ''
                    maxMembri: 0
                    membri:
                      - id: ''
                        nome: ''
                        cognome: ''
                        role: ''
                  destinatario:
                    id: ''
                    nome: ''
                    cognome: ''
                    role: ''
      parameters:
        - name: utenteId
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: 071910ca-5974-43d0-9095-c766a7eb558b
  /api/invites/rifiuta?utenteId=81fd65b0-64e3-4b54-ad49-587c5c6386f2&nomeTeam=piedi:
    post:
      summary: Rifiuta un invito
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_inviti_rifiuta_un_invito_yml
      description: ''
      tags:
        - Inviti
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: utenteId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: 81fd65b0-64e3-4b54-ad49-587c5c6386f2
        - name: nomeTeam
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: piedi
...
servers:
  - url: http://localhost:8080
    description: Server di sviluppo
components:
  schemas:
...
    invia_un_invito_a_un_utente:
      type: object
      properties:
        teamMittenteNome:
          type: string
        destinatarioId:
          type: string
      example:
        teamMittenteNome: piedi
        destinatarioId: dadd5008-e361-480e-95b9-89097d364fd5
...
  requestBodies:
...
    invia_un_invito_a_un_utente:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/invia_un_invito_a_un_utente'
      description: ''
      required: true
...
  securitySchemes: {}